# chore(site): migrate OG image generation to Satori