--- a/site/package.json
+++ b/site/package.json
@@ -41,6 +41,7 @@
     "prism-react-renderer": "^2.4.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "satori": "^0.18.3",
     "sharp": "^0.34.4",
     "typescript": "^5.9.2"
   },