--- a/site/package.json
+++ b/site/package.json
@@ -26,7 +26,6 @@
     "@mdx-js/react": "^3.1.1",
     "@monaco-editor/react": "^4.7.0",
     "@mui/icons-material": "^7.3.2",
-    "@resvg/resvg-js": "^2.6.2",
     "@segment/ajv-human-errors": "^2.15.0",
     "@swc/core": "^1.13.5",
     "@swc/jest": "^0.2.39",