## Summary

Migrates Open Graph (OG) image generation from Resvg to Satori, achieving **2.5x faster image generation** with improved visual quality and smaller file sizes.

## 🎨 Visual Quality

The new Satori implementation produces cleaner, more minimal OG images with:

- ✅ Cleaner left accent bar design (vs full rounded border)
- ✅ Natural typography with proper letter spacing
- ✅ Same Inter SemiBold font and brand colors
- ✅ 60-70% smaller file sizes

**To preview:** Run `node site/src/plugins/docusaurus-plugin-og-image/index.js` to generate sample images in `test-og-output/`

## Performance Improvements

| Metric | Before (Resvg) | After (Satori) | Improvement |
|--------|----------------|----------------|-------------|
| **Text-only images** | 565ms avg | 80ms avg | **7.1x faster** |
| **Images with hero** | 606ms avg | 214ms avg | **2.8x faster** |
| **Batch processing** | Batch size: 2 | Batch size: 8 | **4x more parallel** |
| **File size (text)** | 52-64 KB | 20-28 KB | **~60% smaller** |

**Expected impact on documentation builds:**
- 100 images: ~16 seconds → ~6 seconds (60% faster)
- 500 images: ~80 seconds → ~31 seconds (61% faster)

## What Changed

### Core Implementation
- ✅ Replace Resvg SVG rendering with Satori JSX-based rendering
- ✅ Migrate from string-based SVG templates to typed JSX structures
- ✅ Switch from Resvg PNG conversion to Sharp with compression

### Performance Optimizations
- ✅ **Global font caching**: Load Inter SemiBold font once, reuse across all images
- ✅ **Increased batch size**: From 2 to 8 (Satori has no system font bottleneck)
- ✅ **PNG compression**: Added compressionLevel 9 + palette mode for smaller files
- ✅ **No system fonts**: Eliminates 150-200ms overhead per image from Resvg

### Bug Fixes
- ✅ Fix corrupted `Inter-SemiBold.ttf` font file (was HTML, now proper TTF)

## Technical Benefits

1. **No system font dependency**
   - Resvg requires loading system fonts (~150-200ms per image)
   - Satori uses embedded fonts only
   - More reliable across different environments

2. **Pure JavaScript**
   - No native Rust dependencies to compile
   - Works everywhere Node.js works
   - Easier to maintain and debug

3. **Better parallelization**
   - Increased batch size from 2 to 8
   - Satori's architecture allows better CPU utilization
   - Font loading is cached globally

4. **Modern architecture**
   - JSX-based templates (type-safe, composable)
   - Actively maintained by Vercel
   - Designed specifically for OG images

## Visual Changes

The Satori implementation produces **cleaner, more minimal OG images**:

- **Border style**: Simplified to left accent bar (was full rounded border)
- **Typography**: Natural letter spacing (was stretched with textLength)
- **File size**: ~60% smaller due to PNG compression
- **Quality**: Same Inter SemiBold font, colors, and layout

All existing functionality maintained:
- ✅ Hero images with automatic resizing
- ✅ Breadcrumb navigation
- ✅ Page type badges
- ✅ Metadata extraction from frontmatter
- ✅ HTML meta tag injection

## Testing

- [x] Standalone test runner works (`node site/src/plugins/docusaurus-plugin-og-image/index.js`)
- [x] Visual quality validated with side-by-side comparison
- [x] File sizes verified to be 60-70% smaller
- [x] Performance benchmarked across multiple test cases
- [x] Lint and format checks passed

## Compatibility

- ✅ All existing routes supported
- ✅ Environment variable controls maintained (`SKIP_OG_GENERATION`, `OG_BATCH_SIZE`)
- ✅ Same output file structure
- ✅ Same meta tag injection logic

## Migration Notes

This is a **drop-in replacement** - no configuration changes required:
- Existing OG image URLs remain the same
- Docusaurus plugin lifecycle hooks unchanged
- Environment variables work as before
- Can be reverted easily if needed