@mldangelo 2025-10-04
## 🎨 Visual Quality Comparison

### Before (Resvg) vs After (Satori)

I've generated sample OG images to compare visual quality and file sizes:

#### 1. Short Title Example

<table>
<tr>
<td><b>Before (Resvg)</b><br/>51 KB</td>
<td><b>After (Satori)</b><br/>21 KB (59% smaller)</td>
</tr>
<tr>
<td>

![short-title-before](https://github.com/user-attachments/assets/short-title-resvg.png)

</td>
<td>

![short-title-after](https://github.com/user-attachments/assets/short-title-satori.png)

</td>
</tr>
</table>

#### 2. Long Title Example

<table>
<tr>
<td><b>Before (Resvg)</b><br/>63 KB</td>
<td><b>After (Satori)</b><br/>25 KB (60% smaller)</td>
</tr>
<tr>
<td>

![long-title-before](https://github.com/user-attachments/assets/long-title-resvg.png)

</td>
<td>

![long-title-after](https://github.com/user-attachments/assets/long-title-satori.png)

</td>
</tr>
</table>

#### 3. With Hero Image

<table>
<tr>
<td><b>Before (Resvg)</b><br/>283 KB</td>
<td><b>After (Satori)</b><br/>81 KB (71% smaller)</td>
</tr>
<tr>
<td>

![with-image-before](https://github.com/user-attachments/assets/with-image-resvg.png)

</td>
<td>

![with-image-after](https://github.com/user-attachments/assets/with-image-satori.png)

</td>
</tr>
</table>

### Key Visual Differences

The Satori implementation features a **cleaner, more minimal design**:
- ✅ Simplified left accent bar (vs full rounded border)
- ✅ Natural letter spacing (vs stretched text with `textLength`)
- ✅ Same Inter SemiBold font and brand colors
- ✅ Better readability with proper typography

### File Size Summary

| Type | Before (Resvg) | After (Satori) | Reduction |
|------|----------------|----------------|-----------|
| Short title | 51 KB | 21 KB | **59%** |
| Long title | 63 KB | 25 KB | **60%** |
| With image | 283 KB | 81 KB | **71%** |

**Average reduction: ~63% smaller files** thanks to PNG compression (level 9 + palette mode).